resolved bugs
